requirements.txt.txt
gradio
pytesseract
torch
transformers
Pillow
